% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{AG_test_stat}
\alias{AG_test_stat}
\title{Andersen-Gill test}
\usage{
AG_test_stat(data1_format, data2_format)
}
\arguments{
\item{data1_format}{Data for sample 1 formatted using \code{format_data_AG}}

\item{data2_format}{Data for sample 2 formatted using \code{format_data_AG}}
}
\value{
A \code{list} object which contains
\itemize {
\item{test_stat_p}{p-value for the test}
}
}
\description{
Andersen-Gill model based two-sample test.
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
