% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairtest.R
\name{dtail}
\alias{dtail}
\title{Choose appropriate tail for two-sided test}
\usage{
dtail(stat)
}
\arguments{
\item{stat}{test statistic}
}
\value{
numeric. p-value for the two-sided test
}
\description{
For internal use only. This function returns the correct p-value
for the Wald two-tailed test given the value of the statistic.
}
