% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{format_data_AG}
\alias{format_data_AG}
\title{Format observed data for Andersen-Gill method}
\usage{
format_data_AG(X, delta, Z_star)
}
\arguments{
\item{X}{vector containing observed time to terminating event}

\item{delta}{vector that is 1 if terminating event is observed and 0 if patient is censored}

\item{Z_star}{FILL IN}
}
\value{
A \code{list} object which contains
\itemize {
\item{ID}{Patient ID}
\item{T_start}{start time}
\item{T_stop}{Stop time}
\item{status}{status indicator: 1 if event occured, 0 if censored}
}
}
\description{
Format observed data for Andersen-Gill method
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
