% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{format_data_GL}
\alias{format_data_GL}
\title{Format observed data for Ghosh and Lin method}
\usage{
format_data_GL(X, delta, Z_star, time)
}
\arguments{
\item{X}{vector containing observed time to terminating event}

\item{delta}{vector that is 1 if terminating event is observed and 0 if patient
is censored}

\item{Z_star}{fill}

\item{time}{vector containing start times of follow-up windows chosen}
}
\value{
A \code{list} object which contains
\itemize {
\item{Y}{}
\item{tau}{}
\item{dmu_hat}{}
\item{dch_hat}{}
\item{dpsi_i}{}
\item{dM_D_i}{}
}
}
\description{
Format observed data for Ghosh and Lin method
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
