% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{get_mu_hat_star_tau}
\alias{get_mu_hat_star_tau}
\title{Estimate and variance for Tayob and Murray}
\usage{
get_mu_hat_star_tau(X_km, delta_km, Tau, t)
}
\arguments{
\item{X_km}{vector containing observed time to combined end-point for each follow-up window}

\item{delta_km}{vector containing combined end-point event indicator. 1 if
combined-endpoint is observed and 0 if censored}

\item{Tau}{length of follow-up intervals of interest}

\item{t}{vector containing start times of follow-up windows chosen}
}
\value{
a \code{list} containing
\itemize{
\item{mean}{estimate of overall tau restricted mean survival}
\item{var}{empirical variance estimate of mean}
}
}
\description{
Get estimate and variance of estimate for method described in Tayob, Murray 2015.
This function is used internally when TM() is called.
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
