% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{GL_test_stat}
\alias{GL_test_stat}
\title{Calculate test statistic for Ghosh and Lin method}
\usage{
GL_test_stat(p_GL, time, data1_format, data2_format)
}
\arguments{
\item{p_GL}{p parameter}

\item{time}{time to event}

\item{data1_format}{Formatted data set 1}

\item{data2_format}{Formatted data set 2}
}
\value{
A \code{list} object which contains
\itemize{
\item{test_stat_p}{P-value for test statistic}
}
}
\description{
Calculate test statistic for Ghosh and Lin method
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
