% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched.survtest.diffn.R
\name{matched.survtest.9.10.99}
\alias{matched.survtest.9.10.99}
\title{Title for matched.survtest function. <10 words}
\usage{
matched.survtest.9.10.99(
  x1,
  delta1,
  x2,
  delta2,
  n,
  tm = sort(unique(c(0, x1, x2))),
  weights = "left",
  maxtau = 1e+05,
  fudge.factor = 0
)
}
\arguments{
\item{x1}{insert description and data type, default value if needed}

\item{delta1}{insert description and data type, default value if needed}

\item{x2}{insert description and data type, default value if needed}

\item{delta2}{insert description and data type, default value if needed}

\item{n}{the integer number of correlated individuals and we assume they are
sorted as in the paper, so that correlated observations come first up to the nth 
observation and then uncorrelated observations follow.}

\item{tm}{insert description and data type, default value if needed}

\item{weights}{string, options are "left" and "right". Option "left" will give pf and 
yls weighting using upperlimit similar to logrank. "right" will use weights 
that are similar, but right continuous.}

\item{maxtau}{insert description and data type, default value if needed}

\item{fudge.factor}{insert description and data type, default value if needed}
}
\value{
insert description and data type and what it contains
}
\description{
Short description for function. What does it do? What does the output include?
Is this the main function or a helper?
}
\examples{

}
\references{

}
\seealso{

}
