% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairdata}
\alias{pairdata}
\title{Paired survival data set for examples and tests.}
\format{A \code{data.frame} with 3711 observations (rows) and 6 variables (columns):
\describe{
  \item{trt1}{description, type}
  \item{x1}{description, type}
  \item{delta1}{description, type}
  \item{trt2}{description, type}
  \item{x2}{description, type}
  \item{delta2}{description, type}
}}
\usage{
pairdata
}
\description{
Describe the data set.
}
\keyword{datasets}
