% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairtest.R
\name{pairtest}
\alias{pairtest}
\title{Paired Two-Sample Tests for Survival Endpoints}
\usage{
pairtest(
  x1,
  delta1,
  x2,
  delta2,
  n,
  tm = sort(unique(c(0, x1, x2))),
  weights = "left",
  maxtau = 1e+05
)
}
\arguments{
\item{x1}{numeric, time to event for treatment group 1}

\item{delta1}{numeric, 1 if subject died, 0 if censored}

\item{x2}{numeric, time to event for treatment group 2}

\item{delta2}{numeric, 1 if subject died, 0 if censored}

\item{n}{the integer number of correlated individuals and we assume they are
sorted as in the paper, so that correlated observations come first up to the nth
observation and then uncorrelated observations follow.}

\item{tm}{sorted vector of times to events}

\item{weights}{string, options are "left" and "right". Option "left" will give pf and
yls weighting using upperlimit similar to logrank. "right" will use weights
that are similar, but right continuous.}

\item{maxtau}{numeric, maximum tau cutoff}
}
\value{
a \code{list} object which contains named two-sample statistics and
p-values for the independent and correlated cases, and other auxiliary information
relevant to the analysis.
}
\description{
Perform two-sample tests for treatment effects with paired censored
survival data. The method requires the time to event or a censored observation
and the event indicator for each pair of observations.
}
\examples{
data(pairdata)
eyeresults = pairtest(x1 = pairdata$x1, delta1 = pairdata$delta1, 
                      x2 = pairdata$x2, delta2 = pairdata$delta2, n = 3711)
print(eyeresults)

}
\references{
Murray, Susan. Nonparametric Rank-Based Methods for Group
Sequential Monitoring of Paired Censored Survival Data. 2000. Biometrics,
56, pp. 984-990.
}
\seealso{
TM(), TM2()
}
