% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{plot.TM}
\alias{plot.TM}
\title{Plot a TM object}
\usage{
\method{plot}{TM}(x, ...)
}
\arguments{
\item{x}{an object of class 'TM'}

\item{...}{addtional options to be passed on}
}
\description{
Plot a TM object
}
