% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM2.R
\name{plot.TM2}
\alias{plot.TM2}
\title{Plot a TM2 object}
\usage{
\method{plot}{TM2}(x, ...)
}
\arguments{
\item{x}{an object of class 'TM2'}

\item{...}{additional arguments}
}
\description{
Plot a TM2 object
}
\details{
Four additional plot parameters are available. They include:
\itemize{
\item{alpha - significance level, default is 0.05}
\item{conservative_index - minimum number of events after the start of the last interval,
default is 25}
\item{k - number of points for intergration within follow-up window, default is 500}
\item{n.sim - number of samples simulated to calculate confidence bands, default is 1000}
}
}
