% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairtest.R
\name{print.pairtest}
\alias{print.pairtest}
\title{Print a pairtest object}
\usage{
\method{print}{pairtest}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class 'pairtest'}

\item{digits}{numeric, number of digits to include after the decimal}
}
\description{
Print a pairtest object
}
