% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{print.TM}
\alias{print.TM}
\title{Print a TM object}
\usage{
\method{print}{TM}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class 'TM'}
}
\description{
Print a TM object
}
