% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{RMRL_function}
\alias{RMRL_function}
\title{Calculate RMRL}
\usage{
RMRL_function(data_format, Tau, t)
}
\arguments{
\item{data_format}{output of function "format_data_ourmethod"}

\item{Tau}{length of follow-up intervals of interest}

\item{t}{vector containing start times of follow-up windows chosen}
}
\value{
a \code{list} containing
\itemize{
\item{RMRL - restricted mean residual life function evaluated at times defined by t}
\item{area_under_RMRL - area under RMRL function evaluated at times t}
\item{var_area_under_RMRL - empirical variance of estimate area_under_RMRL}
}
}
\description{
Area under RMRL and variance of area under RMRL.
RMRL is the restricted mean residual life function evaluated at times defined by t.
This function is used internally when TM() is called.
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
