% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairtest.R
\name{round2}
\alias{round2}
\title{Round for print.pairtest}
\usage{
round2(x, ...)
}
\arguments{
\item{x}{test statistic}

\item{...}{other information to be passed to round}
}
\value{
p-value or string summarizing p-value
}
\description{
If p-value is small, return a string,
else return the p-value.
}
