% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{sum_function}
\alias{sum_function}
\title{Helper function for column sums of upper trianglular matrix}
\usage{
sum_function(X)
}
\arguments{
\item{X}{matrix}
}
\description{
Helper function for column sums of upper trianglular matrix
}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
