% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairtest.R
\name{summary.pairtest}
\alias{summary.pairtest}
\title{Print the summary of a pairtest object}
\usage{
\method{summary}{pairtest}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class 'pairtest'}

\item{digits}{number of digits to round to after decimal}

\item{...}{additional options}
}
\description{
Print the summary of a pairtest object
}
