% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM2.R
\name{summary.TM2}
\alias{summary.TM2}
\title{Summarize a TM2 object}
\usage{
\method{summary}{TM2}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class 'TM2'}
}
\description{
Summarize a TM2 object
}
