% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TM2data}
\alias{TM2data}
\title{Repeated time-to-event data with censoring}
\format{A \code{data.frame} with 100 observations (rows) and 2 variables (columns):
\describe{
\item{X}{numeric, time to follow-up or event}
\item{delta}{numeric, 1 if subject died, 0 if censored}
}}
\usage{
TM2data
}
\description{
Example data set for the paper: Tayob, N. and Murray, S., 2016. Nonparametric
restricted mean analysis across multiple follow-up intervals. Statistics & probability
letters, 109, pp.152-158.
}
\keyword{datasets}
