% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMdata}
\alias{TMdata}
\title{Repeated time-to-event data with two treatment groups}
\format{A \code{data.frame} with 200 observations (rows) and 10 variables (columns):
\describe{
\item{Subject_ID}{numeric, unique subject identifier}
\item{X}{numeric, continuous covariate}
\item{delta}{numeric, 1 if subject died, 0 if censored}
\item{Group}{numeric, treatment group}
\item{Z1}{numeric, time to first follow-up}
\item{Z2}{numeric, time to second follow-up}
\item{Z3}{numeric, time to third follow-up}
\item{Z4}{numeric, time to fourth follow-up}
\item{Z5}{numeric, time to fifth follow-up}
}}
\usage{
TMdata
}
\description{
Example data set for the paper: Tayob, N. and Murray, S., 2014. Nonparametric tests
of treatment effect based on combined endpoints for mortality and recurrent events.
Biostatistics, 16(1), pp.73-83.
}
\details{
The data set contains repeated measures time-to-event data with two treatment
groups and a single continuous covariate. Summary of the data set:
}
\keyword{datasets}
