% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{summary.TM}
\alias{summary.TM}
\title{Print the summary of a TM object}
\usage{
\method{summary}{TM}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class 'TM'}
}
\description{
Print the summary of a TM object
}
