% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM.R
\name{TM}
\alias{TM}
\title{Two-sample recurrent events test}
\usage{
TM(X, delta, Z, Group, Tau, t, method = "average")
}
\arguments{
\item{delta}{vector that is 1 if terminating event is observed and 0 if patient is censored}

\item{Z}{array of times to recurrent events. Number of columns corresponds to
maximum number of recurrent events observed for a patient}

\item{Group}{vector indicating which group each patient belongs to}

\item{Tau}{upper limit of integration (cannot be greater than largest follow-up
time, cannot be negative)}

\item{t}{vector containing start times of follow-up windows chosen}

\item{method}{Choose "average" for mean difference test or "area" for area between
the RMRL curves}

\item{x}{vector containing observed time to terminating event}

\item{plot}{TRUE if plots to be generated}
}
\value{
A \code{list} object which contains
\itemize {
\item{Mean}{vector containing sample estimates of overall tau-restricted mean survival in each group}
\item{Var}{vector containing empirical variance of estimates of overall tau-restricted mean survival in each group}
\item{test_stat}{test statistic of two-sample test}
\item{test_stat_p}{p-value of two-sample test}
}
}
\description{
Perform the Tayob, Murray two-sample recurrent events test described in
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\examples{


}
\references{
Tayob, N. and Murray, S., 2014. Nonparametric tests of treatment
effect based on combined endpoints for mortality and recurrent events. Biostatistics,
16(1), pp.73-83.
}
\author{
Nabihah Tayob
}
